test: move new test to dedicated class
